21.09 Implemented "components" field for named routes
