+++ .lintstagedrc
@@ -1,5 +1,4 @@
 {
-    "*": "npm run test",
     "lib/src/**/*.ts": [
         "npm run lint"
     ]
